// generated/src/app/api-models/member-master-edi.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class MemberMasterEdi {
  transactionSetId: string;
  seqMembId: number;
  seqSubsId: number;
  subscriberId: string;
  personNumber: string;
  lastName: string;
  firstName: string;
  middleInitial: string;
  addressLine1: string;
  addressLine2: string;
  city: string;
  state: string;
  zipCode: string;
  homePhoneNumber: string;
  busPhoneNumber: string;
  contactTitle: string;
  dateOfBirth: Date;
  gender: string;
  maritalStatus: string;
  languageCode: string;
  userDefined1: string;
  userDefined2: string;
  medicareNo: string;
  medicaidNo: string;
  socialSecNo: string;
  employeeNo: string;
  medicalRecNo: string;
  seqAltMembId: number;
  holdReason: string;
  holdDate: Date;
  depVerifStatus: string;
  akaLastName: string;
  akaFirstName: string;
  akaAddressLine1: string;
  akaAddressLine2: string;
  akaCity: string;
  akaState: string;
  akaZipCode: string;
  akaPhoneNumber: string;
  respLastName: string;
  respFirstName: string;
  respAddressLine1: string;
  respAddressLine2: string;
  respCity: string;
  respState: string;
  respZipCode: string;
  respPhoneNumber: string;
  verfiedThruDate: Date;
  subsUpdate: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  caseManagementSwitch: string;
  akaCountry: string;
  respCountry: string;
  country: string;
}